refactor(compile): extract import-resolution helpers from collect_modules
collect_modules.rs was 2002 lines — 2 over the lint file-size cap (it
crossed the line during the CI-allowlist outage when the gate wasn't
running). Move the cohesive import-helper cluster (env_defines_for_lowering,
collect_js_module_imports, ResolvedImport + the cached_resolve_*/
source_visible_resolved_path resolvers, known_node_submodule_key) into a
new collect_modules/import_helpers.rs submodule. Pure code-move, no behavior
change; collect_modules.rs is now 1787 lines. 25 collect_modules tests pass.
